HW 7. moved to directories
