+++ HW_5/OOP-functions/oop/Components/css/DivElement.css
@@ -1,5 +1,5 @@
 .widget__icon {
-    background-image: url(icons/Sunny_with_cloud.svg);
+    background-image: url(../icons/Sunny_with_cloud.svg);
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
