+++ HW_5/OOP-functions/oop/Components/css/Input.css
@@ -31,7 +31,7 @@
 
 .task__checkbox:checked {
     background-color: #D3D3D3;
-    background-image: url(icons/Checked.svg);
+    background-image: url(../icons/Checked.svg);
     background-repeat: no-repeat;
     background-position: center;
 }
