+++ HW_5/OOP-functions/oop/Components/css/Label.css
