+++ HW_5/OOP-functions/oop/Interfaces/HTMLCommonElement.ts
