+++ HW_5/OOP-functions/oop/Interfaces/ItemInterface.ts
