+++ HW_5/OOP-functions/oop/Interfaces/Properities.ts
