+++ HW_5/OOP-functions/oop/Interfaces/StateInterface.ts
